mqtt-client/requirements.txt
python-dotenv
psycopg
psycopg2-binary
psycopg-binary
psycopg[binary,pool]
paho-mqtt